+++ roxctl/sensor/generate/generate.go
@@ -29,11 +29,6 @@ be removed in a future version of roxctl. Remove it from your invocations of rox
 versions of roxctl. If you want to preserve the old behavior, please change your
 invocations to explicitly specify --create-upgrader-sa=false.
 `
-
-	warningDefaultForCollectionMethodWillChange = `WARNING: The default for the --collection-method flag will change to "kernel-module" in future
-versions of roxctl. If you want to preserve the old behavior, please change your
-invocations to explicitly specify --collection-method=auto.
-`
 )
 
 var (
@@ -114,12 +109,6 @@ func Command() *cobra.Command {
 			if !c.Flags().Lookup("create-upgrader-sa").Changed {
 				fmt.Fprintf(os.Stderr, "%s\n", warningDefaultForCreateUpgraderSaWillChange)
 			}
-
-			// Migration process for changed default for "--collection-method".
-			// Can be removed in a future release.
-			if !c.Flags().Lookup("collection-method").Changed {
-				fmt.Fprintf(os.Stderr, "%s\n", warningDefaultForCollectionMethodWillChange)
-			}
 		},
 	}
 
@@ -135,7 +124,7 @@ func Command() *cobra.Command {
 
 	c.PersistentFlags().BoolVar(&cluster.RuntimeSupport, "runtime", true, "whether or not to have runtime support (DEPRECATED, use Collection Method instead)")
 
-	c.PersistentFlags().Var(&collectionTypeWrapper{CollectionMethod: &cluster.CollectionMethod}, "collection-method", "which collection method to use for runtime support (none, auto, kernel-module, ebpf)")
+	c.PersistentFlags().Var(&collectionTypeWrapper{CollectionMethod: &cluster.CollectionMethod}, "collection-method", "which collection method to use for runtime support (none, default, kernel-module, ebpf)")
 
 	c.PersistentFlags().BoolVar(&createUpgraderSA, "create-upgrader-sa", false, "whether to create the upgrader service account, with cluster-admin privileges, to facilitate automated sensor upgrades")